# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MiriGangNeung Backend API
  version: 0.0.1
  description: |
    미리강릉 백엔드의 현재 P0 API 계약이다.
    인증 기능은 아직 없으며, 모든 API는 /api/v1 아래에 있다.
servers:
  - url: http://localhost:8080
    description: Local/Docker
tags:
  - name: Places
  - name: Images
  - name: Compositions
  - name: Courses
  - name: Routes
paths:
  /api/v1/places:
    get:
      tags: [Places]
      summary: 관광지 목록 조회 및 키워드 검색
      parameters:
        - {name: category, in: query, required: false, schema: {type: string}}
        - {name: keyword, in: query, required: false, schema: {type: string, default: ''}}
        - {name: page, in: query, required: false, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: 목록 조회 성공
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlacePageResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
  /api/v1/places/{placeId}:
    get:
      tags: [Places]
      summary: 관광지 상세 조회
      parameters:
        - {$ref: '#/components/parameters/PlaceId'}
      responses:
        '200':
          description: 상세 조회 성공
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlaceDetailResponse'}
        '404': {$ref: '#/components/responses/PlaceNotFound'}
  /api/v1/places/{placeId}/nearby:
    get:
      tags: [Places]
      summary: 주변 관광지 조회
      parameters:
        - {$ref: '#/components/parameters/PlaceId'}
      responses:
        '200':
          description: 주변 관광지 목록
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/PlaceResponse'}}
  /api/v1/places/{placeId}/related:
    get:
      tags: [Places]
      summary: 연관 관광지 조회
      parameters:
        - {$ref: '#/components/parameters/PlaceId'}
      responses:
        '200':
          description: 연관 관광지 목록
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/PlaceResponse'}}
  /media/images/{storageKey}:
    get:
      tags: [Images]
      summary: 저장된 장소 이미지 스트리밍
      description: CDN으로 교체 가능한 이미지 origin endpoint다. 성공 응답은 장기 immutable Cache-Control을 제공한다.
      parameters:
        - name: storageKey
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: 이미지 파일
          headers:
            Cache-Control:
              schema: {type: string, example: 'public, max-age=31536000, immutable'}
          content:
            image/*:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/compositions:
    post:
      tags: [Compositions]
      summary: 사진 합성 Job 생성
      description: Place.id UUID와 Type1 관광지 원본을 사용해 Agent generation Job을 생성한다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [onePickId]
              properties:
                photo: {type: string, format: binary}
                modelPresetId:
                  type: string
                  description: 기본 AI 모델 ID. photo와 정확히 하나만 전송한다.
                  example: default-female-01
                onePickId: {type: string, format: uuid}
                aspectRatio: {type: string, enum: ['1:1', '4:5', '9:16'], default: '4:5'}
                backgroundImageUrl:
                  type: string
                  nullable: true
                  description: 해당 Place에서 사용자가 선택한 Type1 원본 이미지 URL. 생략 시 첫 Type1 이미지 사용.
      responses:
        '200':
          description: Job 생성 결과
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CompositionStatusResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '415': {description: 지원하지 않는 이미지 형식}
        '422': {description: 합성 가능한 Type1 배경 없음}
        '502': {description: 배경 원본 또는 Agent 호출 실패}
        '503': {description: AI_BASE_URL 미설정}
  /api/v1/composition-models:
    get:
      tags: [Compositions]
      summary: 기본 AI 모델 목록 조회
      responses:
        '200':
          description: 사용 가능한 기본 AI 모델
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompositionModelResponse'
  /api/v1/composition-models/{modelPresetId}/image:
    get:
      tags: [Compositions]
      summary: 기본 AI 모델 이미지 조회
      parameters:
        - name: modelPresetId
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: 모델 이미지
          content:
            image/png:
              schema: {type: string, format: binary}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '503': {description: 모델 asset을 사용할 수 없음}
  /api/v1/compositions/{jobId}:
    get:
      tags: [Compositions]
      summary: 사진 합성 Job 상태 조회
      parameters:
        - {$ref: '#/components/parameters/JobId'}
      responses:
        '200':
          description: Job 상태
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CompositionStatusResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/compositions/{jobId}/retry:
    post:
      tags: [Compositions]
      summary: 사진 합성 Job 재시도
      parameters:
        - {$ref: '#/components/parameters/JobId'}
      responses:
        '200':
          description: 재시도 결과
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CompositionStatusResponse'}
        '409': {description: 재시도 불가능한 상태 또는 오류}
        '410': {description: 사용자 원본 이미지 TTL 만료}
  /api/v1/compositions/{jobId}/download:
    get:
      tags: [Compositions]
      summary: 합성 결과 다운로드
      parameters:
        - {$ref: '#/components/parameters/JobId'}
      responses:
        '200': {description: 합성 이미지 다운로드, content: {image/*: {schema: {type: string, format: binary}}}}
        '409': {description: 결과가 아직 준비되지 않음}
        '410': {description: 결과 TTL 만료}
  /api/v1/courses:
    post:
      tags: [Courses]
      summary: 규칙 기반 코스 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCourseRequest'}
            example:
              placeIds: ['백엔드에서 받은 place id']
              onePickId: '백엔드에서 받은 place id'
              types: [food, rest, culture, nature]
              detailTypes: [food:chinese, food:japanese, rest:coffee]
              companion: couple
              duration: day
      responses:
        '200':
          description: 코스 생성 성공
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CourseResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
  /api/v1/courses/{courseId}:
    get:
      tags: [Courses]
      summary: 코스 조회
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
      responses:
        '200': {description: 코스 응답, content: {application/json: {schema: {$ref: '#/components/schemas/CourseResponse'}}}}
    delete:
      tags: [Courses]
      summary: 코스 삭제
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
      responses:
        '200': {description: 삭제 성공}
  /api/v1/courses/{courseId}/nearby-places:
    get:
      tags: [Courses]
      summary: 코스 주변 또는 강릉 전체 장소 조회
      description: scope=nearby는 코스 관광지 주변을 추천/거리순으로 조회하며 반경은 사용자 선택 없이 2km → 5km → 10km → 15km로 자동 확장한다. 세부 취향이 있고 현재 반경 안의 정확한 일치 후보가 3개 미만이면 다음 반경을 조회하고, 최종 반경은 응답의 searchRadiusMeters로 안내한다. scope=all은 사용자가 제출한 키워드로 기준 좌표 없는 Gangneung rect 안의 선택 카테고리를 조회한다. scope=all에 keyword가 없으면 외부 API를 호출하지 않고 빈 결과를 반환한다. 강릉 대표 모드는 현재 클라이언트 준비 상태이며 이 API scope로 전송하지 않는다.
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
        - name: scope
          in: query
          required: false
          description: nearby(기본값) 또는 all. 대표 모드는 UI에서 준비 중으로만 표시한다.
          schema: {type: string, enum: [nearby, all], default: nearby}
        - name: category
          in: query
          required: true
          schema: {type: string, enum: [restaurant, cafe, culture, attraction]}
        - name: stopId
          in: query
          required: false
          description: 특정 관광지 정거장 주변만 조회할 때 사용하는 stop UUID
          schema: {type: string, format: uuid}
        - name: sort
          in: query
          required: false
          description: 추천 점수순 또는 거리순. 기본값은 recommended.
          schema: {type: string, enum: [recommended, distance], default: recommended}
        - name: keyword
          in: query
          required: false
          description: scope=all일 때 검색 실행을 위해 제출하는 장소명 키워드. 비어 있으면 빈 결과를 반환한다.
          schema: {type: string}
        - name: page
          in: query
          required: false
          description: 0부터 시작하는 Kakao 결과 페이지(최대 44; Kakao의 1~45페이지에 대응)
          schema: {type: integer, minimum: 0, maximum: 44, default: 0}
        - name: size
          in: query
          required: false
          description: 페이지 크기(최대 15)
          schema: {type: integer, minimum: 1, maximum: 15, default: 15}
      responses:
        '200':
          description: scope에 따른 주변/강릉 전체 장소 페이지
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NearbyPlacesResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '502': {$ref: '#/components/responses/ExternalApiError'}
  /api/v1/courses/{courseId}/stops/external:
    post:
      tags: [Courses]
      summary: Kakao 주변 장소를 코스 snapshot으로 추가
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddExternalStopRequest'}
      responses:
        '200':
          description: 장소가 추가된 코스
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CourseResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '409': {description: 이미 코스에 있는 장소}
  /api/v1/courses/{courseId}/stops/{stopId}:
    delete:
      tags: [Courses]
      summary: 코스 장소 삭제
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
        - name: stopId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: 장소가 삭제된 코스
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CourseResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/courses/{courseId}/stops/order:
    put:
      tags: [Courses]
      summary: 코스 장소 순서 변경
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StopOrderRequest'}
      responses:
        '200':
          description: 순서가 변경된 코스
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CourseResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
  /api/v1/courses/{courseId}/stops/optimize:
    post:
      tags: [Courses]
      summary: Kakao 도보 거리 기준으로 코스 순서 최적화
      description: 관광지와 추가 장소 전체의 방문 순서를 바꿀 수 있다. 후보 순서의 실제 도보 거리 합계가 현재보다 짧을 때만 저장한다.
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
      responses:
        '200':
          description: 최적화된 순서와 도보 경로를 포함한 코스 (개선이 없으면 기존 순서)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CourseResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/courses/{courseId}/share:
    post:
      tags: [Courses]
      summary: 익명 공유 링크 생성
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
      responses:
        '200': {description: 공유 정보, content: {application/json: {schema: {$ref: '#/components/schemas/ShareResponse'}}}}
    delete:
      tags: [Courses]
      summary: 공유 링크 폐기
      parameters:
        - {$ref: '#/components/parameters/CourseId'}
      responses:
        '200': {description: 폐기 성공}
  /api/v1/share/courses/{token}:
    get:
      tags: [Courses]
      summary: 공유 토큰으로 코스 조회
      parameters:
        - {name: token, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: 공유 코스, content: {application/json: {schema: {$ref: '#/components/schemas/CourseResponse'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/routes/walking:
    post:
      tags: [Routes]
      summary: 두 지점 사이 도보 경로 조회
      description: 현재 백엔드 계약은 origin과 destination 두 지점이다.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WalkingRouteRequest'}
      responses:
        '200': {description: 정규화된 경로, content: {application/json: {schema: {$ref: '#/components/schemas/WalkingRouteResponse'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '502': {$ref: '#/components/responses/ExternalApiError'}
components:
  parameters:
    PlaceId: {name: placeId, in: path, required: true, schema: {type: string}}
    JobId: {name: jobId, in: path, required: true, schema: {type: string, format: uuid}}
    CourseId: {name: courseId, in: path, required: true, schema: {type: string, format: uuid}}
  responses:
    InvalidRequest:
      description: 요청값이 올바르지 않음
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: 대상을 찾을 수 없음
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    PlaceNotFound:
      description: 관광지를 찾을 수 없음
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ExternalApiError:
      description: 외부 API 오류
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  schemas:
    PlaceResponse:
      type: object
      required: [id, name, region, category, tags, latitude, longitude, imageUrls, originalImageUrls]
      properties:
        id: {type: string}
        name: {type: string}
        shortDescription: {type: string, nullable: true, description: 백엔드 큐레이션 장소 소개}
        region: {type: string}
        category: {type: string}
        tags: {type: array, items: {type: string}}
        thumbnailUrl: {type: string, nullable: true, description: 카드 대표 썸네일 URL}
        latitude: {type: number, format: double, nullable: true}
        longitude: {type: number, format: double, nullable: true}
        imageUrls:
          type: array
          maxItems: 5
          items: {type: string}
          description: 카드 표시용 썸네일 URL 목록. originalImageUrls와 같은 순서다.
        originalImageUrls:
          type: array
          maxItems: 5
          items: {type: string}
          description: 합성·대형 표시용 원본 URL 목록. imageUrls와 같은 인덱스를 사용한다.
    PlacePageResponse:
      type: object
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/PlaceResponse'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
    PlaceDetailResponse:
      allOf:
        - {$ref: '#/components/schemas/PlaceResponse'}
        - type: object
          properties:
            description: {type: string, nullable: true}
            imageUrls: {type: array, items: {type: string}}
            originalImageUrls: {type: array, items: {type: string}}
            images: {type: array, items: {$ref: '#/components/schemas/PlaceImageResponse'}}
    PlaceImageResponse:
      type: object
      properties:
        imageUrl: {type: string, description: 카드 표시용 썸네일 URL}
        originalImageUrl: {type: string, description: 합성·대형 표시용 원본 URL}
        sourceImageUrl: {type: string, description: 한국관광공사 원본 URL, attribution/debug 용도}
        title: {type: string, nullable: true}
        source: {type: string, nullable: true}
        sortOrder: {type: integer, nullable: true}
        copyrightCode: {type: string, nullable: true, example: Type1}
    CreateCourseRequest:
      type: object
      required: [placeIds, onePickId, types, companion, duration]
      properties:
        placeIds: {type: array, minItems: 1, maxItems: 20, items: {type: string}}
        onePickId: {type: string}
        types: {type: array, minItems: 1, maxItems: 4, items: {type: string}}
        detailTypes: {type: array, items: {type: string}, description: 선택한 broad 여행 타입의 세부 선호 ID(분야별 복수 선택 가능)}
        companion: {type: string}
        duration: {type: string}
        startDate: {type: string, format: date, nullable: true}
        endDate: {type: string, format: date, nullable: true}
    CourseResponse:
      type: object
      properties:
        courseId: {type: string, format: uuid}
        title: {type: string}
        duration: {type: string}
        types: {type: array, items: {type: string}, description: 코스 생성 시 선택한 여행 타입(최대 4개)}
        detailTypes: {type: array, items: {type: string}, description: 코스 생성 시 선택한 세부 선호}
        companion: {type: string, description: 코스 생성 시 선택한 동행 유형}
        stops: {type: array, items: {$ref: '#/components/schemas/CourseStop'}}
        totalDistanceMeters: {type: integer}
        totalTravelMinutes: {type: integer}
        routeStatus: {type: string, enum: [READY, UNAVAILABLE]}
        routeSegments: {type: array, items: {$ref: '#/components/schemas/RouteSegment'}}
    CourseStop:
      type: object
      properties:
        stopId: {type: string, format: uuid, nullable: true}
        sequence: {type: integer}
        placeId: {type: string, nullable: true}
        externalPlaceId: {type: string, nullable: true}
        name: {type: string}
        thumbnailUrl: {type: string, nullable: true}
        arrivalTime: {type: string}
        stayMinutes: {type: integer}
        crowdLevel: {type: string}
        isOnePick: {type: boolean}
        note: {type: string}
        latitude: {type: number, format: double, nullable: true}
        longitude: {type: number, format: double, nullable: true}
        external: {type: boolean}
        category: {type: string, nullable: true}
        categoryName: {type: string, nullable: true}
        address: {type: string, nullable: true}
        phone: {type: string, nullable: true}
        placeUrl: {type: string, nullable: true}
    RouteSegment:
      type: object
      properties:
        fromStopId: {type: string, format: uuid, nullable: true}
        toStopId: {type: string, format: uuid, nullable: true}
        distanceMeters: {type: integer}
        durationSeconds: {type: integer}
        polyline: {type: array, items: {type: array, minItems: 2, maxItems: 2, items: {type: number}}}
    NearbyPlacesResponse:
      type: object
      properties:
        scope: {type: string, enum: [nearby, all]}
        category: {type: string, enum: [restaurant, cafe, culture, attraction]}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 15, description: nearby는 Kakao 페이지 요청 크기이며 places 배열은 여러 관광지·페이지 결과를 합쳐 15개를 초과할 수 있음}
        isEnd: {type: boolean}
        searchRadiusMeters: {type: integer, nullable: true, description: nearby의 마지막 자동 검색 반경. scope=all에서는 null}
        places: {type: array, items: {$ref: '#/components/schemas/NearbyPlace'}}
    NearbyPlace:
      type: object
      properties:
        externalPlaceId: {type: string}
        name: {type: string}
        category: {type: string, enum: [restaurant, cafe, culture, attraction]}
        categoryName: {type: string, nullable: true}
        address: {type: string, nullable: true}
        roadAddress: {type: string, nullable: true}
        phone: {type: string, nullable: true}
        placeUrl: {type: string, nullable: true}
        latitude: {type: number, format: double}
        longitude: {type: number, format: double}
        distanceMeters: {type: integer, nullable: true, description: scope=all에서는 null}
        nearestStopId: {type: string, format: uuid, nullable: true}
        nearestStopName: {type: string, nullable: true}
        recommendationScore: {type: integer, minimum: 0, maximum: 100, nullable: true, description: 코스 선호 기반 추천 점수. 선호값이 없는 구 코스는 null}
        recommendationReasons: {type: array, maxItems: 3, items: {type: string}, description: 추천 점수의 설명. 리뷰·별점이 아님}
    AddExternalStopRequest:
      type: object
      required: [externalPlaceId, name, category, longitude, latitude]
      properties:
        externalPlaceId: {type: string}
        name: {type: string}
        category: {type: string, enum: [restaurant, cafe, attraction, culture]}
        categoryName: {type: string, nullable: true}
        address: {type: string, nullable: true}
        roadAddress: {type: string, nullable: true}
        phone: {type: string, nullable: true}
        placeUrl: {type: string, nullable: true}
        longitude: {type: number, format: double, minimum: 124, maximum: 132}
        latitude: {type: number, format: double, minimum: 33, maximum: 39}
    StopOrderRequest:
      type: object
      required: [stopIds]
      properties:
        stopIds: {type: array, minItems: 1, items: {type: string, format: uuid}}
    ShareResponse:
      type: object
      properties:
        shareToken: {type: string}
        shareUrl: {type: string}
        expiresAt: {type: string, format: date-time}
    CompositionModelResponse:
      type: object
      required: [id, name, imageUrl]
      properties:
        id: {type: string}
        name: {type: string}
        imageUrl: {type: string}
        description: {type: string, nullable: true}
    CompositionStatusResponse:
      type: object
      required: [jobId, status, progress, stage, resultAvailable]
      properties:
        jobId: {type: string, format: uuid}
        status: {type: string, enum: [QUEUED, ANALYZING, COMPOSITING, QUALITY_CHECK, DONE, FAILED]}
        progress: {type: integer}
        stage: {type: string, nullable: true}
        resultAvailable: {type: boolean}
        downloadUrl: {type: string, nullable: true}
        place: {type: object, nullable: true}
        error:
          allOf:
            - {$ref: '#/components/schemas/CompositionError'}
          nullable: true
        safety:
          allOf:
            - {$ref: '#/components/schemas/CompositionSafety'}
          nullable: true
    CompositionError:
      type: object
      required: [code, message, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
    CompositionSafety:
      type: object
      properties:
        status: {type: string, nullable: true, enum: [PASSED, REJECTED, UNKNOWN]}
        reasonCode: {type: string, nullable: true}
        warnings:
          type: array
          items: {$ref: '#/components/schemas/CompositionWarning'}
    CompositionWarning:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
    WalkingRouteRequest:
      type: object
      required: [origin, destination]
      properties:
        origin: {$ref: '#/components/schemas/Coordinate'}
        destination: {$ref: '#/components/schemas/Coordinate'}
    Coordinate:
      type: object
      required: [latitude, longitude]
      properties:
        latitude: {type: number, format: double}
        longitude: {type: number, format: double}
    WalkingRouteResponse:
      type: object
      properties:
        distanceMeters: {type: integer}
        durationSeconds: {type: integer}
        polyline: {type: array, items: {type: array, minItems: 2, maxItems: 2, items: {type: number}}}
    ErrorResponse:
      type: object
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}